MM-22989 - Moving reply count function to utils
